1270번 문제
| Nij | <= 2^31이라는 문제 조건은 생각보다 많은 걸 담고 있었다.
첫째로, 번호가 반드시 양수일 거라는 고정관념에 의해 2번을 내리 틀렸고,
둘째로, int의 범위가 문제였다.
signed의 int는 -2^31 < x < 2^31 - 1이기 때문에, 만약 정답이 2^31이라면, 메모리 부족으로 잘못된 값을 내보낼 수 밖에 없다.
JAVA의 경우에는 메모리 부족을 런타임 에러로 출력하겠지만, C/C++은 대신 쓰레기 값을 내보내므로, 원인을 추측하기 어려웠던 것이다.

+++ Baekjoon/군대에서 문제풀기/1270.cpp
@@ -20,8 +20,8 @@ int main() {
     for(int i = 0; i < n; i++) {
         cin >> ti;
         
-        unordered_map<int, int> milCount;
-        int mil;
+        unordered_map<long long, int> milCount;
+        long long mil;
 		bool isOccupied = false;
         for(int j = 0; j < ti; j++) {
             cin >> mil;
@@ -38,9 +38,9 @@ int main() {
 				}
 			}
         }
-        
-        if(!isOccupied)
-            answer.push_back("SYJKGW");
+
+		if(!isOccupied)
+			answer.push_back("SYJKGW");
     }
     
     for(int i = 0; i < n; i++)
